Add files via upload
TSP describes a scenario where a salesman wishes to visit a list of cities, taking the shortest possible route, before returning home to the start point.
This problem not having known polynomial time solution and also no time-efficient way to prove the given solution.
The difficulty of this problem increase with increase the no. of nodes in the networking.